Rename tof_profile*.h files
