+++ src/dials/algorithms/integration/tof/tof_profile_1d_ibix.h
@@ -1,5 +1,5 @@
-#ifndef DIALS_ALGORITHMS_INTEGRATION_TOF_TOF_PROFILE1D_H
-#define DIALS_ALGORITHMS_INTEGRATION_TOF_TOF_PROFILE1D_H
+#ifndef DIALS_ALGORITHMS_INTEGRATION_TOF_TOF_PROFILE_1D_IBIX_H
+#define DIALS_ALGORITHMS_INTEGRATION_TOF_TOF_PROFILE_1D_IBIX_H
 
 #include <algorithm>
 #include <cmath>
@@ -602,4 +602,4 @@ namespace dials { namespace algorithms {
   }
 
 }}  // namespace dials::algorithms
-#endif /* DIALS_ALGORITHMS_INTEGRATION_TOF_TOF_PROFILE1D */
+#endif /* DIALS_ALGORITHMS_INTEGRATION_TOF_TOF_PROFILE_1D_IBIX_H */
